+++ client/context/query.go
@@ -103,7 +103,7 @@ func (ctx CLIContext) queryABCI(req abci.RequestQuery) (abci.ResponseQuery, erro
 		return result.Response, nil
 	}
 
-	if err = ctx.verifyProof(req.Path, result.Response); err != nil {
+	if err := ctx.verifyProof(req.Path, result.Response); err != nil {
 		return abci.ResponseQuery{}, err
 	}